Moving file filter so that its dependency lies in another package. for osgi reasons
